Configuration refactored
