+++ Mapper.Playground/Models/ActorMapConfiguration.cs
@@ -4,9 +4,8 @@ public class ActorMapConfiguration : MapConfiguration
 {
     protected override void Prepare()
     {
-        ApplyOptions<Actor, ActorModel>(options => options
+        SetOptions<Actor, ActorModel>(options => options
                 .Ignore(_ => _.Name)
-                //.ChangeValueFor(r => r.Name, s => s.Name + "-lol")
         );
     }
 }
