+++ Mapper/Src/Configurations/MapAct.cs
@@ -5,19 +5,11 @@ using SimpleTools.Mapper.Primitivies;
 
 namespace SimpleTools.Mapper;
 
-public class MapAct<TSource, TResult> : IMapAct
+internal class MapAct<TSource, TResult> : IMapAct<TSource, TResult>
 {
     private readonly ICollection<MapCriterion> _сriteria = new List<MapCriterion>();
-
-    public void Do<TResult>(ref TResult result)
-    {
-        foreach (var сriterion in _сriteria)
-        {
-            сriterion.Action(result);
-        }
-    }
-
-    public MapAct<TSource, TResult> Ignore<TMember>(Expression<Func<TSource, TMember>> expression)
+    
+    public IMapAct<TSource, TResult> Ignore<TMember>(Expression<Func<TSource, TMember>> expression)
     {
         var property = expression.Body as MemberExpression;
         var name = property.Member.Name;
@@ -52,5 +44,10 @@ public class MapAct<TSource, TResult> : IMapAct
         Func<TSource, TAlternate> alternate)
     {
         return this;
+    }
+
+    internal ICollection<MapCriterion> TransitCriteria()
+    {
+        return _сriteria;
     }
 }
